Add  test_sparse_merge (#76)
* add db::tests::test_repeated_write_to_same_key


* add test_sparse_merge


* fix bug in new_iter_iterator and ConcatenatedIterator


* rm tracing log


* ignore test_sparse_merge

+++ src/options.rs
@@ -74,7 +74,7 @@ pub struct Options<C: Comparator> {
     // -------------------
     // Parameters that affect compaction:
     /// The max number of levels except L)
-    pub max_levels: u8,
+    pub max_levels: usize,
 
     /// The number of files necessary to trigger an L0 compaction.
     pub l0_compaction_threshold: usize,
